Project Done

+++ EventManagementAppWebAPI/DTOs/Ticket/TicketBookingDto.cs
@@ -12,7 +12,7 @@
         public int EventId { get; set; }
         public string EventTitle { get; set; } = string.Empty;
         public int Quantity { get; set; }
-        public decimal TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         public DateTime BookingDate { get; set; }
     }
 }
